Arrange buildings problem
You are given a number n, which represents the length of a road. The road has n plots on it's each side.The road is to be so planned that there should not be consecutive buildings on either side of the road. You are required to find and print the number of ways in which the buildings can be built on both side of roads.