Paint House Problem
You are given a number n, representing the number of houses. In the next n rows, you are given 3 space separated numbers representing the cost of painting nth house with red or blue or green color. You are required to calculate and print the minimum cost of painting all houses without painting any consecutive house with same color.